move backup/restore to cmd
