+++ src/cmd/index.ts
@@ -2,6 +2,7 @@
 
 import * as mod1 from'./modules/cmd-conf-mod.js';
 import * as mod2 from'./modules/cmd-app-mod.js';
+import * as mod3 from'./modules/cmd-keys-mod.js';
 
 import { createRequire } from "module";
 const require = createRequire(import.meta.url);
@@ -11,6 +12,7 @@ yargs
   //@ts-ignore
   .command(mod1)
   .command(mod2)
+  .command(mod3)
   .demandCommand()
   .help();
 
@@ -20,3 +22,4 @@ yargs
   .then(() => {
     process.exit(0)
   })
+
